♻️ refactor : develop 내부, User와 Member 중복 문제 해결 - User 관련 내용 삭제 - 폴더 위치 재구조화
