--- src/main/java/com/zip/community/platform/domain/user/UserRole.java
@@ -1,9 +1,0 @@
-package com.zip.community.platform.domain.user;
-
-public enum UserRole {
-
-    FIRST,
-    USER,
-    LAWYER,
-    BROKER
-}
